+++ lib/src/widgets/skadi_paginated_list_view.dart
@@ -2,6 +2,28 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:skadi/skadi.dart';
 
+class SkadiListViewFetchOptions {
+  ///Load more data if we reach this offset start from the bottom
+  ///Make sure to calculate with loading widget height
+  final double fetchOffset;
+
+  ///Auto fetch if the scroll view scrollable area is smaller than this value
+  final double autoFetchOffset;
+
+  ///Auto fetch new data if the List isn't scrollable
+  final bool autoFetchOnShortList;
+
+  ///Recursive the auto fetch until List is scrollable
+  final bool recursiveAutoFetch;
+
+  const SkadiListViewFetchOptions({
+    this.fetchOffset = 0.0,
+    this.autoFetchOffset = 0.0,
+    this.autoFetchOnShortList = false,
+    this.recursiveAutoFetch = false,
+  });
+}
+
 class SkadiPaginatedListView extends StatefulWidget {
   ///Normal List view itemCount
   final int itemCount;
@@ -57,11 +79,8 @@ class SkadiPaginatedListView extends StatefulWidget {
   ///A widget that show at the bottom of ListView when there is an error
   final Widget Function()? errorWidget;
 
-  ///Load more data if we reach this offset start from the bottom
-  ///Make sure to calculate with loading widget height
-  final double fetchOffset;
-
-  final bool autoFetchOnShortList;
+  ///options
+  final SkadiListViewFetchOptions fetchOptions;
 
   const SkadiPaginatedListView({
     Key? key,
@@ -69,16 +88,15 @@ class SkadiPaginatedListView extends StatefulWidget {
     required this.itemBuilder,
     required this.dataLoader,
     required this.hasMoreData,
+    this.fetchOptions = const SkadiListViewFetchOptions(),
     this.physics = const ClampingScrollPhysics(),
     this.shrinkWrap = false,
-    this.fetchOffset = 0.0,
     this.loadingWidget = const CircularProgressIndicator(),
     this.padding = const EdgeInsets.symmetric(vertical: 16, horizontal: 12),
     this.scrollDirection = Axis.vertical,
     this.attachProvidedScrollControllerToListView = false,
     this.hasError = false,
     this.reverse = false,
-    this.autoFetchOnShortList = false,
     this.separatorBuilder,
     this.onEmpty,
     this.scrollController,
@@ -91,6 +109,7 @@ class SkadiPaginatedListView extends StatefulWidget {
 class _SkadiPaginatedListViewState extends State<SkadiPaginatedListView> {
   ScrollController? scrollController;
   int loadingState = 0;
+  bool _stopAutoFetch = false;
 
   bool get _isPrimaryScrollView => widget.scrollController == null;
 
@@ -99,14 +118,14 @@ class _SkadiPaginatedListViewState extends State<SkadiPaginatedListView> {
       return;
     }
     double offsetToFetch =
-        controller.position.maxScrollExtent - widget.fetchOffset;
+        controller.position.maxScrollExtent - widget.fetchOptions.fetchOffset;
     if (controller.offset >= offsetToFetch) {
       loadingState += 1;
       onLoadingMoreData();
     }
   }
 
-  void onLoadingMoreData() async {
+  Future<void> onLoadingMoreData() async {
     if (loadingState > 1) return;
     if (widget.hasMoreData) {
       await widget.dataLoader();
@@ -116,25 +135,40 @@ class _SkadiPaginatedListViewState extends State<SkadiPaginatedListView> {
     }
   }
 
+  void _nonPrimaryListener() {
+    scrollListener(widget.scrollController!);
+  }
+
+  void _primaryListener() {
+    scrollListener(scrollController!);
+  }
+
   void initController() {
     if (_isPrimaryScrollView) {
       scrollController = ScrollController();
-      scrollController!.addListener(() => scrollListener(scrollController!));
+      scrollController!.addListener(_primaryListener);
     } else {
-      widget.scrollController
-          ?.addListener(() => scrollListener(widget.scrollController!));
+      widget.scrollController?.addListener(_nonPrimaryListener);
     }
   }
 
   void checkInitialScrollPosition() {
-    if (widget.autoFetchOnShortList) {
-      Future.delayed(const Duration(milliseconds: 200)).then((timeStamp) {
+    if (widget.fetchOptions.autoFetchOnShortList && !_stopAutoFetch) {
+      Future.delayed(const Duration(milliseconds: 100)).then((timeStamp) {
         if (widget.itemCount > 0) {
           double maxExtents = _isPrimaryScrollView
               ? scrollController!.position.maxScrollExtent
               : widget.scrollController!.position.maxScrollExtent;
-          if (maxExtents <= 0 && !widget.hasError) {
-            onLoadingMoreData();
+          if (maxExtents <= widget.fetchOptions.autoFetchOffset &&
+              !widget.hasError &&
+              widget.hasMoreData) {
+            onLoadingMoreData().then((value) {
+              if (widget.fetchOptions.recursiveAutoFetch) {
+                checkInitialScrollPosition();
+              }
+            });
+          } else {
+            _stopAutoFetch = true;
           }
         }
       });
@@ -143,11 +177,10 @@ class _SkadiPaginatedListViewState extends State<SkadiPaginatedListView> {
 
   void removeListener() {
     if (_isPrimaryScrollView) {
-      scrollController!.removeListener(() => scrollListener(scrollController!));
+      scrollController!.removeListener(_primaryListener);
       scrollController?.dispose();
     } else {
-      widget.scrollController
-          ?.removeListener(() => scrollListener(widget.scrollController!));
+      widget.scrollController?.removeListener(_nonPrimaryListener);
     }
   }